refactor(Finished): Blackjack, Bust, Stay Finished로 추상화
