+++ src/main/java/blackjack/domain/state/finished/Blackjack.java
@@ -1,0 +1,10 @@
+package blackjack.domain.state.finished;
+
+import blackjack.domain.participant.Cards;
+
+public final class Blackjack extends Finished {
+
+    public Blackjack(Cards cards) {
+        super(cards);
+    }
+}
