+++ src/main/java/blackjack/domain/state/finished/Bust.java
@@ -1,0 +1,10 @@
+package blackjack.domain.state.finished;
+
+import blackjack.domain.participant.Cards;
+
+public final class Bust extends Finished {
+
+    public Bust(Cards cards) {
+        super(cards);
+    }
+}
