+++ src/main/java/blackjack/domain/state/finished/Stay.java
@@ -1,0 +1,10 @@
+package blackjack.domain.state.finished;
+
+import blackjack.domain.participant.Cards;
+
+public final class Stay extends Finished {
+
+    public Stay(Cards cards) {
+        super(cards);
+    }
+}
